# public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Laravel
  description: ''
  version: 1.0.0
servers:
  -
    url: 'http://ideas-example.test'
paths:
...
  /api/v1/baskets/create:
    post:
      summary: 'Add Basket'
      operationId: addBasket
      description: ''
      parameters:
        -
          in: header
          name: Content-Type
          description: ''
          example: application/json
          schema:
            type: string
        -
          in: header
          name: Accept
          description: ''
          example: application/json
          schema:
            type: string
      responses: {  }
      tags:
        - Endpoints
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                product_id:
                  type: string
                  description: ''
                  example: voluptatem
                quantity:
                  type: integer
                  description: ''
                  example: 17
              required:
                - product_id
                - quantity
      security: []
  /api/v1/baskets/remove:
    post:
      summary: 'Remove Basket'
      operationId: removeBasket
      description: ''
      parameters:
        -
          in: header
          name: Content-Type
          description: ''
          example: application/json
          schema:
            type: string
        -
          in: header
          name: Accept
          description: ''
          example: application/json
          schema:
            type: string
      responses: {  }
      tags:
        - Endpoints
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                basket_id:
                  type: integer
                  description: ''
                  example: 1
              required:
                - basket_id
      security: []
  /api/v1/baskets/list:
    get:
      summary: 'List Basket'
      operationId: listBasket
      description: ''
      parameters:
        -
          in: header
          name: Content-Type
          description: ''
          example: application/json
          schema:
            type: string
        -
          in: header
          name: Accept
          description: ''
          example: application/json
          schema:
            type: string
      responses:
        401:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  success: false
                  error_message: 'Unauthorized access'
                properties:
                  success:
                    type: boolean
                    example: false
                  error_message:
                    type: string
                    example: 'Unauthorized access'
      tags:
        - Endpoints
      security: []
...